==> notebooks/01_data_exploration.ipynb <==
# Import libraries
import pandas as pd  
import numpy as np   
import matplotlib.pyplot as plt 
import seaborn as sns 

# Load the dataset
df=pd.read_csv('../data/Crop_recommendation_ethiopia_real.csv')

# Basic info
print("Shape: ",df.shape)
print("\nFirst 5 rows:")
print(df.head())

print("\nColumn info: ")
print(df.info())

print("\n Basic statistices: ")
print(df.describe())

print("\n Missing values: ")
print(df.isnull().sum())


print("\nUnique crop(target classes): ")
print(df['label'].unique())
print("Number of unique crops: ", df['label'].nunique())

print("\nClass distribution: ")
print(df['label'].value_counts())
